// frontend/src/app/models/Curso.ts
import { Title } from '@angular/platform-browser';

export interface Curso {
    id?: number;
    imagen?: string;
    nombre?: string;
    objetivo?: string;
    introduccion?: string;
    contenido?: string;
    categoria?: string;
    especialista?: string;
    cantidad?: number;
    duracion?: string;
    descripcion?: string;
    precio?: number;
    vid1?: string;
    url1?: string;
    vid2?: string;
    url2?: string;
    vid3?: string;
    url3?: string;
    vid4?: string;
    url4?: string;
    vid5?: string;
    url5?: string;
    vid6?: string;
    url6?: string;
    vid7?: string;
    url7?: string;
    vid8?: string;
    url8?: string;
    vid9?: string;
    url9?: string;
    vid10?: string;
    url10?: string;
    creacion?: Date;
}